feat(cli): add shared ui package (#20)
* Initial plan

* Changes before error encountered

Agent-Logs-Url: https://github.com/omer-ayhan/skillspp/sessions/b6c6df8f-d133-4b64-835c-af0d314a396e

* feat(core): add plugin source install support

* Initial plan

* Changes before error encountered

Agent-Logs-Url: https://github.com/omer-ayhan/skillspp/sessions/b6c6df8f-d133-4b64-835c-af0d314a396e

* feat(core): add plugin source install support

* feat(cli): add interactive plugin install flow

* feat(cli): add shared cli package

* refactor(cli): adopt shared cli package in apps

* docs(cli): document shared cli package boundaries

* fix: cli shared missing paths for core package

---------
