+++ apps/pluginspp-cli/package.json
@@ -21,6 +21,7 @@
     "lint": "eslint src tests --max-warnings 0"
   },
   "dependencies": {
+    "@skillspp/cli-shared": "workspace:*",
     "commander": "^14.0.3",
     "ink": "^6.8.0",
     "picocolors": "^1.1.1",
