+++ apps/skillspp-cli/package.json
@@ -35,6 +35,7 @@
     "lint": "eslint src tests --max-warnings 0"
   },
   "dependencies": {
+    "@skillspp/cli-shared": "workspace:*",
     "commander": "^14.0.3",
     "gray-matter": "^4.0.3",
     "ink": "^6.8.0",
@@ -44,8 +45,8 @@
     "zod": "^3.25.76"
   },
   "devDependencies": {
-    "@skillspp/core": "workspace:*",
     "@skillspp/platform-node": "workspace:*",
+    "@skillspp/core": "workspace:*",
     "@types/node": "^25.4.0",
     "@types/react": "^19.1.13",
     "@vitest/coverage-v8": "^3.2.4",
