+++ apps/skillspp-cli/src/background-runner.test.ts
@@ -1,15 +1,17 @@
 import { describe, expect, it, vi } from "vitest";
 
-const runBackgroundTaskInPlatform = vi.fn(async () => ({ durationMs: 123 }));
+const createBackgroundTaskRunner = vi.fn();
+const runBackgroundTaskImpl = vi.fn(async () => ({ durationMs: 123 }));
+createBackgroundTaskRunner.mockReturnValue(runBackgroundTaskImpl);
 
-vi.mock("@skillspp/platform-node", () => ({
-  runBackgroundTask: runBackgroundTaskInPlatform,
+vi.mock("@skillspp/cli-shared/runtime/background-runner", () => ({
+  createBackgroundTaskRunner,
 }));
 
 const { runBackgroundTask } = await import("./runtime/background-runner");
 
 describe("CLI runtime background-runner adapter @unit", () => {
-  it("delegates execution to platform-node with resolvable executor module @unit", async () => {
+  it("delegates to the shared background-runner factory @unit", async () => {
     const response = await runBackgroundTask(
       {
         kind: "test.blocking",
@@ -25,10 +27,8 @@ describe("CLI runtime background-runner adapter @unit", () => {
       },
     );
 
-    expect(runBackgroundTaskInPlatform).toHaveBeenCalled();
-    const secondArg = (runBackgroundTaskInPlatform.mock.calls[0] as any)?.[1];
-    expect(typeof secondArg?.executorModule).toBe("string");
-    expect(secondArg.executorModule).toContain("background-executor");
+    expect(createBackgroundTaskRunner).toHaveBeenCalledTimes(1);
+    expect(runBackgroundTaskImpl).toHaveBeenCalled();
     expect(response).toEqual({ durationMs: 123 });
   });
 });
