+++ packages/cli-shared/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./command-builder";
+export * from "./interactive";
