+++ packages/cli-shared/tsconfig.json
@@ -1,0 +1,31 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "noEmit": true,
+    "baseUrl": ".",
+    "paths": {
+      "@skillspp/core": ["../core/src/index.ts"],
+      "@skillspp/core/lockfile": ["../core/src/runtime/lockfile.ts"],
+      "@skillspp/core/telemetry": ["../core/src/runtime/telemetry.ts"],
+      "@skillspp/core/policy": ["../core/src/runtime/policy.ts"],
+      "@skillspp/core/agents": ["../core/src/runtime/agents.ts"],
+      "@skillspp/core/source-parser": ["../core/src/sources/source-parser.ts"],
+      "@skillspp/core/source-resolution": [
+        "../core/src/sources/source-resolution.ts"
+      ],
+      "@skillspp/core/skills": ["../core/src/sources/skills.ts"],
+      "@skillspp/core/commands": ["../core/src/contracts/commands.ts"],
+      "@skillspp/core/results": ["../core/src/contracts/results.ts"],
+      "@skillspp/core/errors": ["../core/src/contracts/errors/core-error.ts"],
+      "@skillspp/core/events": [
+        "../core/src/contracts/events/lifecycle-event.ts"
+      ],
+      "@skillspp/core/contracts/*": ["../core/src/contracts/*"],
+      "@skillspp/core/sources/*": ["../core/src/sources/*"],
+      "@skillspp/core/runtime/*": ["../core/src/runtime/*"],
+      "@skillspp/platform-node": ["../platform-node/src/index.ts"]
+    }
+  },
+  "include": ["src/**/*.ts", "src/**/*.tsx", "tests/**/*.ts", "tests/**/*.tsx"]
+}
